refactor: Move code to cpp files
With the new layer libraries code can be moved into cpp files to proper
hide implementation details.
